% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartsimp.R
\name{bartsimp}
\alias{bartsimp}
\title{Spatial BART Model}
\usage{
bartsimp(
  x.train,
  y.train,
  s1,
  s2,
  x.test = matrix(0, 0, 0),
  s1.test = NULL,
  s2.test = NULL,
  theta = 0,
  omega = 1,
  a = 0.5,
  b = 1,
  augment = FALSE,
  rho = NULL,
  xinfo = matrix(0, 0, 0),
  usequants = FALSE,
  cont = FALSE,
  rm.const = TRUE,
  sigest = 0.1,
  sigdf = 3,
  sigquant = 0.9,
  k = 2,
  power = 2,
  base = 0.95,
  sigmaf = NA,
  lambda = NA,
  treeprev = NA,
  fmean = mean(y.train),
  w = rep(1, length(y.train)),
  ntree = 3L,
  numcut = 100L,
  ndpost = 500L,
  nskip = 1L,
  keepevery = 1L,
  nkeeptrain = ndpost,
  nkeeptest = ndpost,
  nkeeptestmean = ndpost,
  nkeeptreedraws = ndpost,
  printevery = 100L,
  transposed = FALSE,
  seed = 99L,
  nwarmup = 500,
  usecoords = TRUE,
  sigest_new = 1,
  range_new = 5,
  sig_m_new = 1,
  rho_0 = 2.4,
  sigmam_0 = 0.55,
  alpha_1 = 0.05,
  alpha_2 = 0.05
)
}
\arguments{
\item{x.train}{Predictor matrix for training data (n × p).}

\item{y.train}{Numeric response vector of length n.}

\item{s1}{Spatial coordinate vector (length n).}

\item{s2}{Spatial coordinate vector (length n).}

\item{x.test}{Optional predictor matrix for test data.}

\item{sigest}{Initial estimate of residual standard deviation.}

\item{sigdf}{Degrees of freedom for sigma prior.}

\item{sigquant}{Quantile used to calibrate sigma prior.}

\item{k}{Tree prior scale parameter.}

\item{power}{Tree depth prior parameter.}

\item{base}{Tree depth prior parameter.}

\item{ntree}{Number of trees in the ensemble.}

\item{numcut}{Number of cut points per predictor.}

\item{ndpost}{Number of posterior draws to keep.}

\item{nskip}{Number of burn-in iterations.}

\item{keepevery}{Thinning interval.}

\item{seed}{Random seed.}

\item{nwarmup}{Number of warmup iterations.}

\item{usecoords}{Logical; whether to use spatial coordinates.}

\item{size}{Cluster sizes.}
}
\value{
An object containing:

\describe{
\item{yhat.train}{Posterior samples for training data}
\item{yhat.test}{Posterior samples for test data}
\item{yhat.train.mean}{Posterior mean predictions for training data}
\item{yhat.test.mean}{Posterior mean predictions for test data}
\item{sigma}{Posterior draws of the observation-level residual standard deviation}
\item{kappas}{Posterior draws of the Matérn inverse range parameter
\eqn{\kappa = \sqrt{8\nu}/\rho}}
\item{sigmams}{Posterior draws of the spatial random-effect standard deviation}
\item{varcount}{Variable inclusion counts}
\item{varprob}{Variable inclusion probabilities}
\item{proc.time}{Execution time}
}
}
\description{
Fits a Bayesian Additive Regression Trees (BART) model with
spatial variance components.
}
\details{
This function is a simplified interface to the underlying C++ sampler.
The design is inspired by the original BART R package.
}
\examples{
\dontrun{
data("toy_data")

x_train <- toy_data[1:20, c("x1", "x2", "x3", "x4", "x5")]
y_train <- toy_data$y[1:20]

fit <- bartsimp(
  x.train = x_train,
  y.train = y_train,
  s1 = toy_data$s1[1:20],
  s2 = toy_data$s2[1:20],
  ntree = 5,
  ndpost = 5,
  nskip = 2,
  nwarmup = 5,
  seed = 1
)

str(fit$yhat.train.mean)
}

}
\references{
Jiang, A. Z., & Wakefield, J. (2025).
BARTSIMP: Flexible spatial covariate modeling and prediction using
Bayesian Additive Regression Trees.
\emph{Spatial and Spatio-temporal Epidemiology}, 55, 100757.
https://doi.org/10.1016/j.sste.2025.100757

The interface design is inspired by the BART R package.
}
