% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mySPDE}
\alias{mySPDE}
\title{A prototype function that runs the inla function and
returns the marginal log-likelihood}
\usage{
mySPDE(s1, s2, y)
}
\arguments{
\item{s1}{a vector containing the longitude data}

\item{s2}{a vector containing the latitude data}

\item{y}{the response variable}
}
\value{
inla_results the marginal log likelihood

The marginal likelihood for the model
}
\description{
A prototype function that runs the inla function and
returns the marginal log-likelihood
}
\examples{
\dontrun{
s1 <- c(0, 0.5, 1)
s2 <- c(0, 0.5, 1)
y <- c(1.2, 0.8, 1.5)
mySPDE(s1, s2, y)
}

}
