% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{timesTwo}
\alias{timesTwo}
\title{Multiply a number by two}
\usage{
timesTwo()
}
\value{
Integer value \code{0}, invisibly printing diagnostic information from
the native tree test routine.
}
\description{
Multiply a number by two
}
\examples{
timesTwo()

}
