% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-toy_data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Toy dataset for BARTSIMP examples}
\format{
A data frame with spatial replicates.
}
\source{
Simulated data
}
\usage{
toy_data
}
\description{
A small simulated dataset containing:
\itemize{
\item y: response
\item x1-x5: predictors
\item s1, s2: spatial coordinates
\item size: replicate count per location
}
}
\keyword{datasets}
