% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuneprior.R
\name{tunehyperpriors}
\alias{tunehyperpriors}
\title{Tune Spatial Hyperpriors}
\usage{
tunehyperpriors(x_train, y_train, s1, s2)
}
\arguments{
\item{x_train}{Matrix or data frame of training covariates.}

\item{y_train}{Numeric vector of training responses.}

\item{s1}{Numeric vector of first spatial coordinates.}

\item{s2}{Numeric vector of second spatial coordinates.}
}
\value{
A list containing tuned values for \code{sigest_new}, \code{range_new},
\code{sig_m_new}, \code{sigmam_0}, and \code{rho_0}.
}
\description{
Fits a small INLA/SPDE model to obtain rough starting values for the
observation noise, spatial range, and spatial standard deviation priors.
}
\examples{
\dontrun{
data("toy_data")

priors <- tunehyperpriors(
  x_train = toy_data[1:30, c("x1", "x2", "x3", "x4", "x5")],
  y_train = toy_data$y[1:30],
  s1 = toy_data$s1[1:30],
  s2 = toy_data$s2[1:30]
)

priors$sigest_new
}

}
